Fix for the assertion failure when TRACEMEM isn't defined.
I found the cause of the assertion failures I was seeing when
UPPER() and LOWER() were called. The implementation of those
two builtins wasn't allocating enough bytes if the destination
was longer than the source length. Fixed the bug and removed
the FIXME comment from the descrip.mms.

This bug isn't VMS-specific, so I assume nobody caught it
because everyone's compiling with -DNDEBUG to remove the
assertion checks that caught these string overflows.