OpenVMS header file update and more code cleanups.
Updated "vms.h" to reflect that this version of Regina now requires
at least OpenVMS V7.3-2 (__CRTL_VER >= 70312000) to build and run.
It now fails with a preprocessor error if you attempt to build on
anything older. This minimum version requirement is due to the use
of poll() and snprintf(). It also no longer builds on VAX.

Cleaned up "descrip.mms" to build with "NDEBUG" by default, and to
build with warnings enabled, or with /PLUS_LIST for the LTO build.
We also have to use /OPT=(LEV=4) to avoid running out of memory,
but only for /PLUS_LIST builds. Added architecture-specific flags.
The default for Alpha is "/ARCH=EV56/OPT=(TUN=EV67)" due to the
desire to take advantage of EV56 byte-word instructions, and the
lack of benefit of setting /ARCH to anything higher than that.

More minor code cleanups, mostly removing duplicated includes.