Speed up execution by shrinking default hashtable size.
The default hashtable size in variable.c is 2003 entries, which
leads to a hotspot in assign_foliage() when the entire table
has to be iterated over. By reducing the default size to 227 entries,
it's sufficient for the typical Rexx program and empirically speeds
up the RexxCPS benchmark by about 16-18% on my Alpha and my Xeon.