Switch to VMS-style waitpid() for return status.
Using lib$spawn() to spawn DCL commands now requires us to
get the return code using the VMS-style version of getpid(),
by defining _VMS_WAIT. Added _VMS_WAIT #define to vms.h.

Added VMS-specific code to os_unx.c to interpret VMS statuses:
success statuses are checked for the C facility type, in which
case the message number is the POSIX return code. Otherwise,
return 0 for success and 1 for all non-success codes.

I'm currently investigating how to flush the output from DCL
lib$spawn() commands so that Regina can parse it. I think it may be
necessary to write an EOF to the SYS$OUTPUT/SYS$ERROR mailbox when
the subprocess exits, which I'll try to do from an AST in vmscmd.c.

I'll need to store the mailbox name used for SYS$OUTPUT/SYS$ERROR
in the TSD and then write the EOF, or flush it, from the AST.
Hopefully it won't be necessary to modify os_unx.c or shell.c.