Add AST to write EOF to redirected output pipe on completion.
When lib$spawn() is called and the output is redirected to a pipe,
pass the address of an AST function to receive the file descriptor
of the write end of the pipe (which we assign to -1 so that the
caller won't close it immediately) when the subprocess completes.
The AST calls decc$write_eof_to_mbx() and then closes the fd.

With this change, the first few test cases in trip/vms.rexx now
succeed, and I can begin to work on fixes for the bugs that I've
introduced in vmsfuncs.c with my refactoring and modernization.